requirements.txt
google-genai>=1.63.0
python-telegram-bot>=22.6
python-dotenv
flask>=3.0.0
apscheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.0